# docs/swagger.yaml
basePath: /
definitions:
...
  services.AddToCartInput:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  services.UpdateCartInput:
    properties:
      quantity:
        type: integer
    type: object
  types.ChangePasswordInput:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  types.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  types.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  types.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  types.ResetPasswordInput:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - code
    - email
    - new_password
    type: object
host: localhost:8080
info:
  contact:
    email: you@example.com
    name: Developer
  description: API cho hệ thống bán hàng máy tính
  license:
    name: MIT
  termsOfService: http://localhost:8080/
  title: Computer Store API
  version: "1.0"
paths:
...
  /cart:
    get:
      description: Lấy danh sách sản phẩm trong giỏ hàng của người dùng hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lấy giỏ hàng của người dùng
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Người dùng thêm một sản phẩm vào giỏ hàng của họ
      parameters:
      - description: Dữ liệu sản phẩm cần thêm
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.AddToCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Thêm sản phẩm vào giỏ hàng
      tags:
      - Cart
  /cart/{product_id}:
    delete:
      description: Xóa một sản phẩm ra khỏi giỏ hàng của người dùng
      parameters:
      - description: ID sản phẩm
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Xóa sản phẩm khỏi giỏ hàng
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Cập nhật số lượng của một sản phẩm trong giỏ hàng theo product_id
      parameters:
      - description: ID sản phẩm
        in: path
        name: product_id
        required: true
        type: integer
      - description: Số lượng mới
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật số lượng sản phẩm trong giỏ hàng
      tags:
      - Cart
...
  /change-password:
    post:
      consumes:
      - application/json
      description: Người dùng đổi mật khẩu khi đã đăng nhập
      parameters:
      - description: Mật khẩu cũ và mới
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Đổi mật khẩu
      tags:
      - Auth
  /delete-account:
    delete:
      description: Xóa tài khoản người dùng hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Xóa tài khoản
      tags:
      - Auth
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Gửi mã khôi phục đến email
      parameters:
      - description: Email người dùng
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Quên mật khẩu
      tags:
      - Auth
  /login:
    post:
      consumes:
      - application/json
      description: Đăng nhập và nhận JWT token
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Đăng nhập
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Đăng ký người dùng mới
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Đăng ký tài khoản
      tags:
      - Auth
  /reset-password:
    post:
      consumes:
      - application/json
      description: Đặt lại mật khẩu bằng mã khôi phục
      parameters:
      - description: Email, mã code và mật khẩu mới
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Đặt lại mật khẩu
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"